feat: optimized vram usage by 1) filtering only resid_post activations and 2) casting model to dtype float16 instead of transformerlens default 32 (note: llama and qwen use float16 anyways

+++ probe/src/look_ahead_probe/build_look_ahead_activation_dataset.py
@@ -34,7 +34,9 @@ def main():
     args = parser.parse_args()
 
     print(f"Loading model: {args.model_name}")
-    model = HookedTransformer.from_pretrained(args.model_name, device=args.device)
+    model = HookedTransformer.from_pretrained(
+        args.model_name, device=args.device, dtype=torch.bfloat16
+    )
 
     if args.layers is not None:
         layers = [int(x.strip()) for x in args.layers.split(',')]
